Move CSVWriter into its own module
